requirements.txt
asgiref==3.11.1
black==26.3.1
click==8.3.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
psycopg2-binary==2.9.11
python-decouple==3.8
pytokens==0.4.1
sqlparse==0.5.5
